Check if bounce sync is scheduled within 42 hours for all emails when sending
[MAILPOET-2705]
